#402 Changes after review
