+++ accesscontroltool-bundle/src/main/java/biz/netcentric/cq/tools/actool/ui/WebConsoleConfigTracker.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.Dictionary;
 
 import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.jackrabbit.oak.commons.PropertiesUtil;
 import org.osgi.service.cm.ConfigurationAdmin;
 import org.osgi.service.cm.ConfigurationEvent;
@@ -21,20 +20,14 @@ public class WebConsoleConfigTracker implements ConfigurationListener {
 
     private static final Logger LOG = LoggerFactory.getLogger(WebConsoleConfigTracker.class);
 
-    private static final String CONSOLE_PID = "org.apache.felix.webconsole.internal.servlet.OsgiManager";
-    private static final String CONSOLE_ROOT_PROP = "manager.root";
-    private static final String CONSOLE_ROOT_DEFAULT = "/system/console"; 
-    
     private static final String CONSOLE_SEC_PROVIDER_PID = "org.apache.sling.extensions.webconsolesecurityprovider.internal.SlingWebConsoleSecurityProvider";
     private static final String CONSOLE_SEC_PROVIDER_USERS_PROP = "users";
     private static final String CONSOLE_SEC_PROVIDER_GROUPS_PROP = "groups";
 
-    private static final String[] RELEVANT_PIDS = new String[] {CONSOLE_PID, CONSOLE_SEC_PROVIDER_PID};
+    private static final String[] RELEVANT_PIDS = new String[] {CONSOLE_SEC_PROVIDER_PID};
 
     @Reference(policyOption = ReferencePolicyOption.GREEDY)
     private ConfigurationAdmin configAdmin;
-
-    private String webConsoleRoot;
 
     private String[] allowedUsers = new String[] {};
     private String[] allowedGroups = new String[] {};
@@ -42,11 +35,6 @@ public class WebConsoleConfigTracker implements ConfigurationListener {
     @Activate
     private void updateConfig() {
         try {
-            Dictionary<String, Object> webconsoleConfig = configAdmin.getConfiguration(CONSOLE_PID).getProperties();
-            if(webconsoleConfig != null) {
-                webConsoleRoot = (String) webconsoleConfig.get(CONSOLE_ROOT_PROP);
-            }
-            webConsoleRoot = StringUtils.defaultIfBlank(webConsoleRoot, CONSOLE_ROOT_DEFAULT);
 
             Dictionary<String, Object> webconsoleSecProviderConfig = configAdmin.getConfiguration(CONSOLE_SEC_PROVIDER_PID).getProperties();
             if(webconsoleSecProviderConfig != null) {
@@ -54,8 +42,7 @@ public class WebConsoleConfigTracker implements ConfigurationListener {
                 allowedGroups = PropertiesUtil.toStringArray(webconsoleSecProviderConfig.get(CONSOLE_SEC_PROVIDER_GROUPS_PROP));
             }
             if(LOG.isDebugEnabled()) {
-                LOG.debug("webConsoleRoot: {} allowedUsers: {} allowedGroups: {}", 
-                        webConsoleRoot, 
+                LOG.debug("allowedUsers: {} allowedGroups: {}", 
                         ArrayUtils.toString(allowedUsers), 
                         ArrayUtils.toString(allowedGroups));
             }
@@ -71,10 +58,6 @@ public class WebConsoleConfigTracker implements ConfigurationListener {
         if(ArrayUtils.contains(RELEVANT_PIDS, pid)) {
             updateConfig();
         }
-    }
-
-    public String getWebConsoleRoot() {
-        return webConsoleRoot;
     }
 
     public String[] getAllowedUsers() {
